Move bootstrap to internal/xds
