+++ internal/xds/bootstrap/template.go
