+++ internal/xds/bootstrap/template_test.go
