+++ xds/internal/xdsclient/clientimpl_authority.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 
 	"google.golang.org/grpc/internal/grpclog"
-	"google.golang.org/grpc/xds/internal/xdsclient/bootstrap"
+	"google.golang.org/grpc/internal/xds/bootstrap"
 	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource"
 )
