+++ xds/internal/xdsclient/xdsresource/resource_type.go
@@ -28,8 +28,8 @@ import (
 	"fmt"
 
 	"google.golang.org/grpc/internal"
+	"google.golang.org/grpc/internal/xds/bootstrap"
 	xdsinternal "google.golang.org/grpc/xds/internal"
-	"google.golang.org/grpc/xds/internal/xdsclient/bootstrap"
 	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource/version"
 	"google.golang.org/protobuf/types/known/anypb"
 )
